Introduce Morgan, a PyPI mirror for restricted networks
Morgan is a PyPI mirror for restricted networks with no access to
the Internet, and potentially multiple Python environments with
different versions of CPython, different operating systems, etc.

This initial version contains an advanced mirrorer for usage in
unrestricted networks that generates a directory tree (a "package
index") with both source and binary distributions of specific
requirements, and a single-file server application. This tree can be
copied/rsynced to the target restricted environments, going through
whatever security practices are in place. The server file can then be
executed by whatever CPython interpreter available (so long as it is
version 3.7 or above).

Read the README.md file for more information.