Replace pkg_resources with importlib.metadata
The pkg_resources library from setuptools - which was added in
commit 34a86aa123 to generate list of requirements from the current
environment - documents that its usage is discouraged in favor of
importlib.metadata, which is included in Python's standard library
starting in version 3.8, and also includes a backport called
importlib_metadata for older versions.

This commit replaces pkg_resources with importlib.metadata in Python
versions 3.8 and above, and with importlib_metadata in lower versions
(only 3.7 is officially supported by morgan).

A bug is also fixed for older versions of Python: type hints need to use
object classes from the typing library instead of directly refering to
classes such as `list`, `tuple` and `dict`.

This is version 0.10.0 of the software.