Move to Hatch, add metadata
The project no longer uses Poetry for a project manager, and now uses
Hatch >=1.5.0. This allows the project to define metadata in
pyproject.toml in a standard way.

A lot of metadata is introduced. The version is no longer statically
defined in pyproject.toml, but dynamically defined in
morgan/__about__.py (variable __version__). A "version" command is added
to the Morgan command line, which simply prints the version and exits.