Add server flag to disable serving of metadata files
By default, the Morgan server implements PEP 658 ("Serve Distribution
Metadata in the Simple Repository API"). This allows clients such as
`pip` to download and parse metadata files without having to download
complete package archives first. However, if the `pip` user provides
non-normalized names to the command (e.g. `pip install FastAPI`
instead of `pip install fastapi`), the install will fail, because the
name inside the returned metadata file is different than the name
provided to `pip`. This seems like a bug in `pip` (because it is
sending normalized names to the server), but it's better to at least
have a workaround.

The server now accepts the `--no-metadata` flag, which will disable
serving of metadata files completely if provided. This allows `pip`
to install with non-normalized names as well.