Bugfix: downloaded files may not match all environments
When Morgan mirrors packages, it chooses the most recent version
that satisfies the requirement strings, and the environments.
However, when matching environments, it would select files (and
therefore version) that matches at least one of the Python versions
of the environments, which could lead to Morgan choosing versions
that do not support all of the Python versions.

This commit fixes this bug by requiring the selected versions match
ALL of the Python versions defined in the environments. This could
potentially lead to Morgan not finding matches at all, but this
would be fixed in a subsequent release by treating each environment
independently of the others.